requirements.txt
albumentations>=1.3.1
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
segmentation-models-pytorch>=0.3.0
PyYAML>=6.0
scikit-image>=0.21.0
